prep: separate CheerioIssueRepository to get timeline for new view
